// @stargazezone/contracts/names/Sg721Name.types.ts
export interface Metadata {
  image_nft?: NFT | null;
  records: TextRecord[];
}
export interface NFT {
  collection: Addr;
  token_id: string;
}
export interface TextRecord {
  name: string;
  value: string;
  verified?: boolean | null;
}
// ...
export interface NftInfoResponseForMetadata {
  extension: Metadata;
  token_uri?: string | null;
}